fix: adjusting spelling
