+++ src/components/inputs/MultipleChoices/stories/props.ts
