+++ src/components/inputs/SelectCheck/OptionItem/index.tsx
@@ -1,32 +1,31 @@
-// import { Label, Stack } from "@inube/design-system";
 import { Label } from "@inubekit/label";
 import { Stack } from "@inubekit/stack";
 
 import { StyledOptionItemChecked } from "./styles";
 
 export interface IOptionItemCheckedProps {
-    id: string;
-    label: string;
-    checked?: boolean;
-    onchange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  id: string;
+  label: string;
+  checked?: boolean;
+  onchange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
 export const OptionItemChecked = (props: IOptionItemCheckedProps) => {
-    const { id, label, checked = false, onchange } = props;
+  const { id, label, checked = false, onchange } = props;
 
-    return (
-        <StyledOptionItemChecked>
-            <Stack>
-                <input
-                    readOnly
-                    type="checkbox"
-                    id={id}
-                    name={id}
-                    onChange={onchange}
-                    checked={checked}
-                />
-                <Label htmlFor={id}>{label}</Label>
-            </Stack>
-        </StyledOptionItemChecked>
-    );
+  return (
+    <StyledOptionItemChecked>
+      <Stack>
+        <input
+          readOnly
+          type="checkbox"
+          id={id}
+          name={id}
+          onChange={onchange}
+          checked={checked}
+        />
+        <Label htmlFor={id}>{label}</Label>
+      </Stack>
+    </StyledOptionItemChecked>
+  );
 };
