+++ src/components/inputs/SelectCheck/stories/SelectCheck.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { ControllerSelectCheck } from "./constroller";
+import { ControllerSelectCheck } from "./controller";
 import { parameters, props } from "./props";
-import { Selectcheck } from "..";
+import { SelectCheck } from "..";
 
 const options = [
   { id: "uno", label: "Option 1", checked: false },
@@ -10,9 +10,9 @@ const options = [
   { id: "tres", label: "Option 3", checked: true },
 ];
 
-type Story = StoryObj<typeof Selectcheck>;
+type Story = StoryObj<typeof SelectCheck>;
 
-const meta: Meta<typeof Selectcheck> = {
+const meta: Meta<typeof SelectCheck> = {
   title: "components/inputs/SelectCheck",
   component: ControllerSelectCheck,
   parameters,
@@ -21,14 +21,14 @@ const meta: Meta<typeof Selectcheck> = {
 
 export const Default: Story = {
   args: {
-    name: "Selectcheck",
-    id: "Selectcheck",
-    placeholder: "Selectcheck",
+    name: "SelectCheck",
+    id: "SelectCheck",
+    placeholder: "SelectCheck",
     disabled: false,
     readonly: false,
     required: false,
     status: "pending",
-    message: "Selectcheck",
+    message: "SelectCheck",
     fullwidth: false,
     options: options,
     onBlur: () => {},
